+++ src/test-targets/metadata.g.ts
@@ -2377,6 +2377,68 @@ export const ComplexModel = domain.types.ComplexModel = {
         role: "value",
       },
     },
+    downloadAttachment_VaryStringCustomCache: {
+      name: "downloadAttachment_VaryStringCustomCache",
+      displayName: "Download Attachment _ Vary String Custom Cache",
+      transportType: "item",
+      httpMethod: "GET",
+      params: {
+        id: {
+          name: "id",
+          displayName: "Primary Key",
+          type: "number",
+          role: "value",
+          get source() { return (domain.types.ComplexModel as ModelType & { name: "ComplexModel" }).props.complexModelId },
+          rules: {
+            required: val => val != null || "Primary Key is required.",
+          }
+        },
+        etag: {
+          name: "etag",
+          displayName: "Etag",
+          type: "string",
+          role: "value",
+          get source() { return (domain.types.ComplexModel as ModelType & { name: "ComplexModel" }).props.name },
+        },
+      },
+      return: {
+        name: "$return",
+        displayName: "Result",
+        type: "file",
+        role: "value",
+      },
+    },
+    downloadAttachment_VaryStringNoCache: {
+      name: "downloadAttachment_VaryStringNoCache",
+      displayName: "Download Attachment _ Vary String No Cache",
+      transportType: "item",
+      httpMethod: "GET",
+      params: {
+        id: {
+          name: "id",
+          displayName: "Primary Key",
+          type: "number",
+          role: "value",
+          get source() { return (domain.types.ComplexModel as ModelType & { name: "ComplexModel" }).props.complexModelId },
+          rules: {
+            required: val => val != null || "Primary Key is required.",
+          }
+        },
+        etag: {
+          name: "etag",
+          displayName: "Etag",
+          type: "string",
+          role: "value",
+          get source() { return (domain.types.ComplexModel as ModelType & { name: "ComplexModel" }).props.name },
+        },
+      },
+      return: {
+        name: "$return",
+        displayName: "Result",
+        type: "file",
+        role: "value",
+      },
+    },
     downloadAttachment_VaryInt: {
       name: "downloadAttachment_VaryInt",
       displayName: "Download Attachment _ Vary Int",