+++ src/test-targets/viewmodels.g.ts
@@ -648,6 +648,28 @@ export class ComplexModelViewModel extends ViewModel<$models.ComplexModel, $apiC
     return downloadAttachment_VaryString
   }
   
+  public get downloadAttachment_VaryStringCustomCache() {
+    const downloadAttachment_VaryStringCustomCache = this.$apiClient.$makeCaller(
+      this.$metadata.methods.downloadAttachment_VaryStringCustomCache,
+      (c) => c.downloadAttachment_VaryStringCustomCache(this.$primaryKey, this.name),
+      () => ({}),
+      (c, args) => c.downloadAttachment_VaryStringCustomCache(this.$primaryKey, this.name))
+    
+    Object.defineProperty(this, 'downloadAttachment_VaryStringCustomCache', {value: downloadAttachment_VaryStringCustomCache});
+    return downloadAttachment_VaryStringCustomCache
+  }
+  
+  public get downloadAttachment_VaryStringNoCache() {
+    const downloadAttachment_VaryStringNoCache = this.$apiClient.$makeCaller(
+      this.$metadata.methods.downloadAttachment_VaryStringNoCache,
+      (c) => c.downloadAttachment_VaryStringNoCache(this.$primaryKey, this.name),
+      () => ({}),
+      (c, args) => c.downloadAttachment_VaryStringNoCache(this.$primaryKey, this.name))
+    
+    Object.defineProperty(this, 'downloadAttachment_VaryStringNoCache', {value: downloadAttachment_VaryStringNoCache});
+    return downloadAttachment_VaryStringNoCache
+  }
+  
   public get downloadAttachment_VaryInt() {
     const downloadAttachment_VaryInt = this.$apiClient.$makeCaller(
       this.$metadata.methods.downloadAttachment_VaryInt,